dir refactor monorepo
