+++ Athos.sln
@@ -1,0 +1,99 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{827E0CD3-B72D-47B6-A68D-7590B98EB39B}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ReviewAutomation", "ReviewAutomation", "{E5034465-0CF5-5EFC-09A4-73D2793BA240}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Api", "Api", "{ED6BEB8B-A9B3-ED17-C698-CBB9173A991A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Athos.ReviewAutomation.Api", "src\ReviewAutomation\Api\Athos.ReviewAutomation.Api.csproj", "{10C24E5D-719E-41B9-A6DE-601E19FB5492}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{9B5C8DEC-00B5-BECC-155A-B2CC9BB1C822}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Athos.ReviewAutomation.Core", "src\ReviewAutomation\Core\Athos.ReviewAutomation.Core.csproj", "{8E228918-2A01-4733-9C14-A0B6B6C8975B}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Infrastructure", "Infrastructure", "{65F71F0C-7182-8BC6-41C6-9FB7860532DD}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Athos.ReviewAutomation.Infrastructure", "src\ReviewAutomation\Infrastructure\Athos.ReviewAutomation.Infrastructure.csproj", "{46BDBE23-7B5A-4BD5-9D6A-7A69C5E55B53}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Models", "Models", "{4346A6AC-958D-2898-265F-CF2F88A0C00F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Athos.ReviewAutomation.Models", "src\ReviewAutomation\Models\Athos.ReviewAutomation.Models.csproj", "{B8EF4D36-2816-4C9F-959A-D60581FA1386}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{10C24E5D-719E-41B9-A6DE-601E19FB5492}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{10C24E5D-719E-41B9-A6DE-601E19FB5492}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{10C24E5D-719E-41B9-A6DE-601E19FB5492}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{10C24E5D-719E-41B9-A6DE-601E19FB5492}.Debug|x64.Build.0 = Debug|Any CPU
+		{10C24E5D-719E-41B9-A6DE-601E19FB5492}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{10C24E5D-719E-41B9-A6DE-601E19FB5492}.Debug|x86.Build.0 = Debug|Any CPU
+		{10C24E5D-719E-41B9-A6DE-601E19FB5492}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{10C24E5D-719E-41B9-A6DE-601E19FB5492}.Release|Any CPU.Build.0 = Release|Any CPU
+		{10C24E5D-719E-41B9-A6DE-601E19FB5492}.Release|x64.ActiveCfg = Release|Any CPU
+		{10C24E5D-719E-41B9-A6DE-601E19FB5492}.Release|x64.Build.0 = Release|Any CPU
+		{10C24E5D-719E-41B9-A6DE-601E19FB5492}.Release|x86.ActiveCfg = Release|Any CPU
+		{10C24E5D-719E-41B9-A6DE-601E19FB5492}.Release|x86.Build.0 = Release|Any CPU
+		{8E228918-2A01-4733-9C14-A0B6B6C8975B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8E228918-2A01-4733-9C14-A0B6B6C8975B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8E228918-2A01-4733-9C14-A0B6B6C8975B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{8E228918-2A01-4733-9C14-A0B6B6C8975B}.Debug|x64.Build.0 = Debug|Any CPU
+		{8E228918-2A01-4733-9C14-A0B6B6C8975B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8E228918-2A01-4733-9C14-A0B6B6C8975B}.Debug|x86.Build.0 = Debug|Any CPU
+		{8E228918-2A01-4733-9C14-A0B6B6C8975B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8E228918-2A01-4733-9C14-A0B6B6C8975B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8E228918-2A01-4733-9C14-A0B6B6C8975B}.Release|x64.ActiveCfg = Release|Any CPU
+		{8E228918-2A01-4733-9C14-A0B6B6C8975B}.Release|x64.Build.0 = Release|Any CPU
+		{8E228918-2A01-4733-9C14-A0B6B6C8975B}.Release|x86.ActiveCfg = Release|Any CPU
+		{8E228918-2A01-4733-9C14-A0B6B6C8975B}.Release|x86.Build.0 = Release|Any CPU
+		{46BDBE23-7B5A-4BD5-9D6A-7A69C5E55B53}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{46BDBE23-7B5A-4BD5-9D6A-7A69C5E55B53}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{46BDBE23-7B5A-4BD5-9D6A-7A69C5E55B53}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{46BDBE23-7B5A-4BD5-9D6A-7A69C5E55B53}.Debug|x64.Build.0 = Debug|Any CPU
+		{46BDBE23-7B5A-4BD5-9D6A-7A69C5E55B53}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{46BDBE23-7B5A-4BD5-9D6A-7A69C5E55B53}.Debug|x86.Build.0 = Debug|Any CPU
+		{46BDBE23-7B5A-4BD5-9D6A-7A69C5E55B53}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{46BDBE23-7B5A-4BD5-9D6A-7A69C5E55B53}.Release|Any CPU.Build.0 = Release|Any CPU
+		{46BDBE23-7B5A-4BD5-9D6A-7A69C5E55B53}.Release|x64.ActiveCfg = Release|Any CPU
+		{46BDBE23-7B5A-4BD5-9D6A-7A69C5E55B53}.Release|x64.Build.0 = Release|Any CPU
+		{46BDBE23-7B5A-4BD5-9D6A-7A69C5E55B53}.Release|x86.ActiveCfg = Release|Any CPU
+		{46BDBE23-7B5A-4BD5-9D6A-7A69C5E55B53}.Release|x86.Build.0 = Release|Any CPU
+		{B8EF4D36-2816-4C9F-959A-D60581FA1386}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B8EF4D36-2816-4C9F-959A-D60581FA1386}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B8EF4D36-2816-4C9F-959A-D60581FA1386}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B8EF4D36-2816-4C9F-959A-D60581FA1386}.Debug|x64.Build.0 = Debug|Any CPU
+		{B8EF4D36-2816-4C9F-959A-D60581FA1386}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B8EF4D36-2816-4C9F-959A-D60581FA1386}.Debug|x86.Build.0 = Debug|Any CPU
+		{B8EF4D36-2816-4C9F-959A-D60581FA1386}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B8EF4D36-2816-4C9F-959A-D60581FA1386}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B8EF4D36-2816-4C9F-959A-D60581FA1386}.Release|x64.ActiveCfg = Release|Any CPU
+		{B8EF4D36-2816-4C9F-959A-D60581FA1386}.Release|x64.Build.0 = Release|Any CPU
+		{B8EF4D36-2816-4C9F-959A-D60581FA1386}.Release|x86.ActiveCfg = Release|Any CPU
+		{B8EF4D36-2816-4C9F-959A-D60581FA1386}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{E5034465-0CF5-5EFC-09A4-73D2793BA240} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{ED6BEB8B-A9B3-ED17-C698-CBB9173A991A} = {E5034465-0CF5-5EFC-09A4-73D2793BA240}
+		{10C24E5D-719E-41B9-A6DE-601E19FB5492} = {ED6BEB8B-A9B3-ED17-C698-CBB9173A991A}
+		{9B5C8DEC-00B5-BECC-155A-B2CC9BB1C822} = {E5034465-0CF5-5EFC-09A4-73D2793BA240}
+		{8E228918-2A01-4733-9C14-A0B6B6C8975B} = {9B5C8DEC-00B5-BECC-155A-B2CC9BB1C822}
+		{65F71F0C-7182-8BC6-41C6-9FB7860532DD} = {E5034465-0CF5-5EFC-09A4-73D2793BA240}
+		{46BDBE23-7B5A-4BD5-9D6A-7A69C5E55B53} = {65F71F0C-7182-8BC6-41C6-9FB7860532DD}
+		{4346A6AC-958D-2898-265F-CF2F88A0C00F} = {E5034465-0CF5-5EFC-09A4-73D2793BA240}
+		{B8EF4D36-2816-4C9F-959A-D60581FA1386} = {4346A6AC-958D-2898-265F-CF2F88A0C00F}
+	EndGlobalSection
+EndGlobal
