--- Models/ReviewResponseDto.cs
@@ -1,8 +1,0 @@
-namespace Athos.Api.Models
-{
-    public class ReviewResponseDto
-    {
-        public string ReviewId { get; set; } = string.Empty;
-        public string FinalResponse { get; set; } = string.Empty;
-    }
-}
