--- Services/NotificationService.cs
@@ -1,16 +1,0 @@
-using Athos.Api.Models;
-
-namespace Athos.Api.Services
-{
-    public class NotificationService
-    {
-        public void SendAlert(DbReview review)
-        {
-            Console.WriteLine("🚨 ALERT: Negative review received");
-            Console.WriteLine($"- Author: {review.Author}");
-            Console.WriteLine($"- Comment: {review.Comment}");
-            Console.WriteLine($"- Rating: {review.Rating}");
-            Console.WriteLine($"- Timestamp: {review.Timestamp}");
-        }
-    }
-}
