+++ src/ReviewAutomation/Api/Athos.ReviewAutomation.Api.csproj
@@ -1,0 +1,25 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+  <PropertyGroup>
+    <TargetFramework>net9.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Core\Athos.ReviewAutomation.Core.csproj" />
+    <ProjectReference Include="..\Infrastructure\Athos.ReviewAutomation.Infrastructure.csproj" />
+    <ProjectReference Include="..\Models\Athos.ReviewAutomation.Models.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="9.0.4" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.4"/>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Content Include="..\Infrastructure\Data\mockGoogleReviews.json">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+  </ItemGroup>
+</Project>
