+++ src/ReviewAutomation/Core/Athos.ReviewAutomation.Core.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+    <PropertyGroup>
+        <TargetFramework>net9.0</TargetFramework>
+        <Nullable>enable</Nullable>
+        <ImplicitUsings>enable</ImplicitUsings>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\Models\Athos.ReviewAutomation.Models.csproj" />
+    </ItemGroup>
+</Project>
