+++ src/ReviewAutomation/Core/Services/AutoReplyService.cs
@@ -1,0 +1,12 @@
+namespace Athos.ReviewAutomation.Core.Services
+{
+    public class AutoReplyService
+    {
+        public string? GenerateReply(string sentiment, string? authorName)
+        {
+            if (sentiment != "Positive") return null;
+
+            return $"Thanks for the kind words{(string.IsNullOrWhiteSpace(authorName) ? "!" : $", {authorName}!")}";
+        }
+    }
+}
