+++ src/ReviewAutomation/Infrastructure/Services/NotificationService.cs
@@ -1,0 +1,17 @@
+using Athos.ReviewAutomation.Models;
+
+
+namespace Athos.ReviewAutomation.Infrastructure.Services
+{
+    public class NotificationService
+    {
+        public void SendAlert(DbReview review)
+        {
+            Console.WriteLine("🚨 ALERT: Negative review received");
+            Console.WriteLine($"- Author: {review.Author}");
+            Console.WriteLine($"- Comment: {review.Comment}");
+            Console.WriteLine($"- Rating: {review.Rating}");
+            Console.WriteLine($"- Timestamp: {review.Timestamp}");
+        }
+    }
+}
