+++ src/ReviewAutomation/Models/ReviewResponseDto.cs
@@ -1,0 +1,8 @@
+namespace Athos.ReviewAutomation.Models
+{
+    public class ReviewResponseDto
+    {
+        public string ReviewId { get; set; } = string.Empty;
+        public string FinalResponse { get; set; } = string.Empty;
+    }
+}
